# openapi.yaml
openapi: 3.0.3
info:
  title: Delivery Service API
  version: 1.0.0
  description: |
    OpenAPI specification generated from existing Spring Boot controllers and DTOs.
    Includes only routes found in code under `src/main/kotlin`.
...
paths:
...
  /api/v1/cart:
    get:
      tags: [Cart]
      operationId: getCart
      summary: Get or create active cart for current actor
      security:
        - bearerAuth: []
        - installIdHeader: []
      responses:
        '200':
          description: Active cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Cart]
      operationId: clearCart
      summary: Remove all items from active cart
      security:
        - bearerAuth: []
        - installIdHeader: []
      responses:
        '200':
          description: Cleared cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/cart/items:
    post:
      tags: [Cart]
      operationId: addCartItem
      summary: Add product (or product variant) with optional modifiers to active cart
      security:
        - bearerAuth: []
        - installIdHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemRequest'
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/cart/items/{itemId}:
    patch:
      tags: [Cart]
      operationId: changeCartItemQuantity
      summary: Change quantity of a cart item
      security:
        - bearerAuth: []
        - installIdHeader: []
      parameters:
        - $ref: '#/components/parameters/CartItemIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeCartItemQuantityRequest'
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Cart]
      operationId: removeCartItem
      summary: Remove one configured cart item from active cart
      security:
        - bearerAuth: []
        - installIdHeader: []
      parameters:
        - $ref: '#/components/parameters/CartItemIdPathParam'
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/catalog/categories:
    get:
      tags: [Catalog]
      operationId: getCategories
      summary: List catalog categories
      description: Returns categories sorted by name ascending.
      parameters:
        - name: activeOnly
          in: query
          required: false
          description: When true (default), returns only active categories.
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/catalog/products:
    get:
      tags: [Catalog]
      operationId: getProducts
      summary: List catalog products
      description: |
        Returns active products sorted by `createdAt` descending.
        Supports filters by `categoryId` and case-insensitive title `query`.
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: query
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/admin/catalog/categories:
    get:
      tags: [Catalog Admin]
      operationId: getAdminCategories
      summary: List categories by activity flag
      security:
        - bearerAuth: []
      parameters:
        - name: isActive
          in: query
          required: true
          description: Returns only categories with the provided activity status.
          schema:
            type: boolean
      responses:
        '200':
          description: Filtered category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Catalog Admin]
      operationId: upsertCategory
      summary: Create or update category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCategoryRequest'
      responses:
        '200':
          description: Saved category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/catalog/products:
    get:
      tags: [Catalog Admin]
      operationId: getAdminProducts
      summary: List products by activity flag
      security:
        - bearerAuth: []
      parameters:
        - name: isActive
          in: query
          required: true
          description: Returns only products with the provided activity status.
          schema:
            type: boolean
      responses:
        '200':
          description: Filtered product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Catalog Admin]
      operationId: upsertProduct
      summary: Create or update product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertProductRequest'
      responses:
        '200':
          description: Saved product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token in `Authorization: Bearer <token>`"
    installIdHeader:
      type: apiKey
      in: header
      name: X-Install-Id
      description: Guest actor identifier used by `CurrentActorArgumentResolver`
  parameters:
...
    CartItemIdPathParam:
      name: itemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequestError:
      description: Request validation failed or business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnauthorizedError:
      description: Authentication required or invalid token/header
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ForbiddenError:
      description: Authenticated but not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, INTERNAL_ERROR]
        message:
          type: string
        traceId:
          type: string
          nullable: true
          description: Request trace id (`X-Trace-Id`) when available.
        details:
          type: object
          nullable: true
          additionalProperties: true
          # TODO: concrete map shape varies by exception type; only stable contract is object map.
...
    AddCartItemRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
          format: uuid
        variantId:
          type: string
          format: uuid
          nullable: true
        quantity:
          type: integer
          format: int32
          minimum: 1
        modifiers:
          type: array
          items:
            $ref: '#/components/schemas/AddCartItemModifierRequest'
    AddCartItemModifierRequest:
      type: object
      required: [modifierGroupId, modifierOptionId]
      properties:
        modifierGroupId:
          type: string
          format: uuid
        modifierOptionId:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
          minimum: 1
          default: 1
    ChangeCartItemQuantityRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          format: int32
          minimum: 1
...
    CartResponse:
      type: object
      required: [id, status, items, totalPriceMinor]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/CartStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItemResponse'
        totalPriceMinor:
          type: integer
          format: int64
        delivery:
          allOf:
            - $ref: '#/components/schemas/CartDeliveryDraftResponse'
          nullable: true
    CartDeliveryDraftResponse:
      type: object
      required: [quoteExpired]
      properties:
        deliveryMethod:
          $ref: '#/components/schemas/DeliveryMethodType'
          nullable: true
        address:
          allOf:
            - $ref: '#/components/schemas/DeliveryAddressResponse'
          nullable: true
        pickupPointId:
          type: string
          format: uuid
          nullable: true
        pickupPointExternalId:
          type: string
          nullable: true
        pickupPointName:
          type: string
          nullable: true
        pickupPointAddress:
          type: string
          nullable: true
        quote:
          allOf:
            - $ref: '#/components/schemas/DeliveryQuoteResponse'
          nullable: true
        quoteExpired:
          type: boolean
        updatedAt:
          type: string
          format: date-time
          nullable: true
    CartItemResponse:
      type: object
      required:
        [id, productId, title, unit, countStep, quantity, priceMinor, unitPriceMinor, modifiersTotalMinor, lineTotalMinor, modifiers]
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        variantId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        unit:
          $ref: '#/components/schemas/ProductUnit'
        countStep:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        priceMinor:
          type: integer
          format: int64
        unitPriceMinor:
          type: integer
          format: int64
        modifiersTotalMinor:
          type: integer
          format: int64
        lineTotalMinor:
          type: integer
          format: int64
        modifiers:
          type: array
          items:
            $ref: '#/components/schemas/CartItemModifierResponse'
    CartItemModifierResponse:
      type: object
      required: [modifierGroupId, modifierOptionId, groupCode, groupName, optionCode, optionName, applicationScope, priceMinor, quantity]
      properties:
        modifierGroupId:
          type: string
          format: uuid
        modifierOptionId:
          type: string
          format: uuid
        groupCode:
          type: string
        groupName:
          type: string
        optionCode:
          type: string
        optionName:
          type: string
        applicationScope:
          $ref: '#/components/schemas/ModifierApplicationScope'
        priceMinor:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
...
    DeliveryQuoteResponse:
      type: object
      required: [deliveryMethod, available, currency]
      properties:
        deliveryMethod:
          $ref: '#/components/schemas/DeliveryMethodType'
        available:
          type: boolean
        priceMinor:
          type: integer
          format: int64
          nullable: true
        currency:
          type: string
        zoneCode:
          type: string
          nullable: true
        zoneName:
          type: string
          nullable: true
        estimatedDays:
          type: integer
          format: int32
          nullable: true
        message:
          type: string
          nullable: true
        pickupPointId:
          type: string
          format: uuid
          nullable: true
        pickupPointExternalId:
          type: string
          nullable: true
        pickupPointName:
          type: string
          nullable: true
        pickupPointAddress:
          type: string
          nullable: true
    CategoryResponse:
      type: object
      required: [id, name, slug, imageUrls, isActive]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        isActive:
          type: boolean
    ProductResponse:
      type: object
      required: [id, categoryId, title, slug, priceMinor, imageUrls, unit, countStep, isActive]
      properties:
        id:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        brand:
          type: string
          nullable: true
        priceMinor:
          type: integer
          format: int64
        oldPriceMinor:
          type: integer
          format: int64
          nullable: true
        sku:
          type: string
          nullable: true
        imageUrls:
          type: array
          items:
            type: string
        unit:
          $ref: '#/components/schemas/ProductUnit'
        countStep:
          type: integer
          format: int32
        isActive:
          type: boolean
...
    UpsertCategoryRequest:
      type: object
      required: [name]
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          minLength: 1
        slug:
          type: string
          nullable: true
        imageIds:
          type: array
          items:
            type: string
            format: uuid
        isActive:
          type: boolean
          default: true
    UpsertProductRequest:
      type: object
      required: [categoryId, title, priceMinor, unit, countStep]
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        categoryId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
        slug:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        priceMinor:
          type: integer
          format: int64
          minimum: 0
        oldPriceMinor:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        sku:
          type: string
          nullable: true
        imageIds:
          type: array
          items:
            type: string
            format: uuid
        unit:
          $ref: '#/components/schemas/ProductUnit'
        countStep:
          type: integer
          format: int32
          minimum: 1
        isActive:
          type: boolean
          default: true
        optionGroups:
          type: array
          items:
            $ref: '#/components/schemas/UpsertProductOptionGroupRequest'
        modifierGroups:
          type: array
          items:
            $ref: '#/components/schemas/UpsertProductModifierGroupLinkRequest'
        variants:
          type: array
          items:
            $ref: '#/components/schemas/UpsertProductVariantRequest'
...
    UpsertProductModifierGroupLinkRequest:
      type: object
      required: [modifierGroupId]
      properties:
        modifierGroupId:
          type: string
          format: uuid
        sortOrder:
          type: integer
          format: int32
          default: 0
        isActive:
          type: boolean
          default: true
    UpsertProductOptionGroupRequest:
      type: object
      required: [code, title]
      properties:
        code:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        sortOrder:
          type: integer
          format: int32
          default: 0
        values:
          type: array
          items:
            $ref: '#/components/schemas/UpsertProductOptionValueRequest'
    UpsertProductOptionValueRequest:
      type: object
      required: [code, title]
      properties:
        code:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        sortOrder:
          type: integer
          format: int32
          default: 0
    UpsertProductVariantRequest:
      type: object
      required: [sku]
      properties:
        externalId:
          type: string
          nullable: true
        sku:
          type: string
          minLength: 1
        title:
          type: string
          nullable: true
        priceMinor:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        oldPriceMinor:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        imageIds:
          type: array
          items:
            type: string
            format: uuid
        sortOrder:
          type: integer
          format: int32
          default: 0
        isActive:
          type: boolean
          default: true
        options:
          type: array
          items:
            $ref: '#/components/schemas/UpsertProductVariantOptionRequest'
    UpsertProductVariantOptionRequest:
      type: object
      required: [optionGroupCode, optionValueCode]
      properties:
        optionGroupCode:
          type: string
          minLength: 1
        optionValueCode:
          type: string
          minLength: 1
...
    DeliveryAddressResponse:
      type: object
      properties:
        country:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        street:
          type: string
          nullable: true
        house:
          type: string
          nullable: true
        apartment:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        entrance:
          type: string
          nullable: true
        floor:
          type: string
          nullable: true
        intercom:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
...
    ModifierApplicationScope:
      type: string
      enum: [PER_ITEM, PER_LINE]
...
    ProductUnit:
      type: string
      enum: [PIECE, KILOGRAM, GRAM, LITER, MILLILITER]
    DeliveryMethodType:
      type: string
      enum: [PICKUP, COURIER, YANDEX_PICKUP_POINT]
...
    CartStatus:
      type: string
      enum: [ACTIVE, ORDERED, MERGED, ABANDONED]
...